changed Item.weight to double in subclasses. rough implementation of Gear.useEnchantment. added enchantment INFERNO. added enchantment attributes muffled & damageNearby.

+++ src/main/java/inte/grupp5/item/Consumable.java
@@ -3,7 +3,7 @@ package inte.grupp5.item;
 // makes further expansion easier.
 public abstract class Consumable extends Item {
 
-    public Consumable(String name, int weight) {
+    public Consumable(String name, double weight) {
         super(name, weight);
     }
 }
